requirements.txt
google-cloud-bigquery
flask
functions-framework==3.4.0
